requirements.txt
Django==5.2.1
djangorestframework==3.14.0
dj-database-url==2.1.0
psycopg2-binary==2.9.11
whitenoise==6.12.0
gunicorn==21.2.0
pillow==12.0.0
python-dotenv==1.2.2
pandas==2.2.2
openpyxl==3.1.2